Relay length attribute mangled by unpack/repack
Great catch from plcp!

  https://trac.torproject.org/projects/tor/ticket/26060

When getting a circuit response we unpack then repack to get the data we just
read from the socket. This is silly. We only did this to separate the cell
headers from the payload but in doing so we corrupted the relay cell's length
attribute.

Better to just read the raw encrypted data and only construct RelayCells
*after* it has been decrypted.