Fix bug in unpacking CREATE_FAST and CREATED_FAST cells due to ZERO stripping
Random bytes can end in b'\x00', so the bug would happen 1/256th of the
time, assuming even random distribution. Only the last byte of the
payload, prior to padding, would affect this.

Additionally, make the implementation ignore the padding portion, rather
than require it must be ZERO.

Furthermore, although DESTROY cells did not suffer from the bug due to a
default content of 0 if fully stripped, standardize the implementation
for them to follow a similar pattern as other cells, relying on Size
pop() for the Exception.

Lastly, adjust the tests for DESTROY cells to accommodate a
non-identical cell due to attempt to recreate it from its payload
(ignoring the padding).