Enforce VPaddingCell constructor specification of payload or size
The behavior of a default range for VPADDING cells is not specified in
the spec, and shouldn't be facilitated by stem.client.

This change also removes a potentially network-visible use of insecure
random generation.

For stem.client randomness, we use secure generation via os.urandom()

(Finally, this removes import of the random module, since it is no
longer used.)