Fix bug that ignored VERSIONS cell link_protocol when packing
The link_protocol was hardcoded to 2 within pack() for VERSIONS but the
spec explicitly states "To be interpreted correctly, later VERSIONS
cells MUST have a CIRCID_LEN matching the version negotiated with the
first VERSIONS cell." (section 4.1).

So while the use of VERSIONS cells after the negotiation is not
expected, stem.client still needs to pack them correctly in case it is
leveraged to test this property in any tor implementation.

The use of link_protocol 2 to set the CIRCID_LEN appropriately has moved
instead to stem.client.Relay, for sending the first VERSIONS cell.