Unpacking RELAY cells now checks data is the specified length
Realistically, since RELAY cells are fixed-width, this was only a
problem for cells which advertise a data length beyond (currently) 498
bytes, which is the max size that the data portion can be.

However, instead of hardcoding that assumption in, the implementation
checks that the specified length of data is indeed unpacked. This should
be more flexible, if ever the format changes in the future.