Use base Cell class for test_unimplemented_cell_methods
No need to subclass. Making a child doesn't hurt, but it also doesn't provide
anything over just instantiating our abstract class.

Also dropping the '^%s$' matching from our regex. This isn't wrong at all - in
fact, it's a bit better. But it's also something applicable to ~80% of the
raise assertions we make. Probably a quirk of mine, but kinda like to avoid the
extra verbosity.

It's kinda questionable if we should be making these assertions at all. The
following makes a case against it...

  https://stackoverflow.com/questions/8672754/how-to-show-the-error-messages-caught-by-assertraises-in-unittest-in-python2-7/8673096#8673096

Maybe that's why the unittest module doesn't provide it.