Extend circ_id validation even further
At its root the prior commit contended with the fact we defaulted circ_id to
zero. Here's the conundrum...

  * Zero is not a valid circuit identifier so using it as a default for
    CircuitCell subclasses is incorrect.

  * We chose zero because *non-curcuit cells* still need to fill something in,
    and zero is a good filler value.

It took me a minute to realize this so clearly the code needs to be clarified.
Hopefully this makes things clearer.

As for the test it's definitely preferable to exercise public methods,  but if
it adds much overhead I don't mind exercising a private method or two. This is
our own tests after all - in java terms we can treat these as 'protected'
rather than fully being off limits. As such, making our test a bit more
targeted while also covering the new checks.