Support vote's new bandwidth-file fields
Support for a newly added network status vote field...

  https://gitweb.torproject.org/torspec.git/commit/?id=84591df

While whipping this up it crossed my mind that the spec is uncommonly loose for
this one. Some examples of odd valid content are...

  * Empty lines: 'bandwidth-file'
  * Multiple equals: 'bandwidth-file timestamp=12=34'
  * Inclusion of these lines in the consensus.

Pointed these out to juga. Also, imho 'headers' should be in the name somewhere
since that's what these are (not a path or file content itself).