'async' SyntaxError with python 3.6 and above
Python 3.6 adds a couple new keywords to the language (async and await) so we
need to be careful not to use those. 99.9% of our users don't import our test
utils so this was mostly moot, but Juan ran into an issue packaging Stem for
Fedora...

  BUILDSTDERR:   File "/usr/lib/python3.7/site-packages/stem/util/test_tools.py", line 151
  BUILDSTDERR:     self.method.async = self
  BUILDSTDERR:                     ^
  BUILDSTDERR: SyntaxError: invalid syntax

  https://trac.torproject.org/projects/tor/ticket/26739

This line aimed to provide a way for tests to reference its AsyncTest instance.
This doesn't look to actually be used in practice so think we can simply drop
it. Checked with grep and a 'run_tests.py --all --target ONLINE' run.