Make LinkProtocol an integer subclass
Link protocols are ints. I made a LinkProtocol class so we can bundle
protocol characteristics (field size and such) but at the end of the
day it's nice for LinkProtocol to still behave like an int.

Dave made the good point that LinkProtocols should be comparable with
ints...

  https://trac.torproject.org/projects/tor/ticket/26432

Taking that one step further and making them in int subclass instead.
Experimenting, the __init__ and __new__ methods seem to work in practice
for our purposes but the python docs advise __new__ so using that...

  https://jfine-python-classes.readthedocs.io/en/latest/subclass-int.html