Add abstract BaseRelayCell class
It's intended to defined common Relay cell functionality, for encrypted
and decrypted cells.

By unpacking content no further than the payload, this abstraction
should allow a lot of flexibility in handling Relay cells prior to
decryption.

Note that unpacking is not yet possible - subclasses must be defined
with a VALUE. (Namely: RELAY and RELAY_EARLY)