Provide None from get_exit_policy when not a relay
When not configured to be a tor relay tor's 'GETINFO exit-policy/full' query
won't function. Rather than attempt to guess the exit policy from other
configuration options providing None for 552 responses...

  https://trac.torproject.org/projects/tor/ticket/25853
  https://gitweb.torproject.org/torspec.git/commit/?id=c5453a0

The above spec addition makes it ambiguous if 552 means 'not a relay' or
'we had an error' so reached out for clarificaition on how these should
be handled...

  https://trac.torproject.org/projects/tor/ticket/27034

In the process also had to fix get_info to raise OperationFailed rather than
a ProtocolError so callers can get the underlying response codes.