ControlMessage equality and hashing
Interesting! Previously I implemented comparability for a single event type but
that's all. I'm surprised we've gone this long without needing it.

Messages have only three things we need to account for in the hash...

  * The message content (ie. the bytes we got from tor).
  * The message class, if we've casted ourselves to something.
  * The arrived_at timestamp if this is an event.

That's it. Everything else is derived from those.

Changing hash_attr()'s parent argument back to a class because a boolean makes
it pretty likely we'll hit an invinite loop. In this case we hit it because we
had the following object hierarchy...

  ControlMessage  <= has a hash method
  Event           <= has a hash method
  BandwidthEvent  <= our object type

If our Event class' hash method calls super it will return the Event class, not
ControlMessage, causing us to loop indefinitely. Changing hash_attr() so it
takes an explicit reference for our parent instead.