Lingering thread due to recent event unit tests
Oops, new tests for #27053 didn't close the threads of the mock controller it
created. Unit tests passed because we don't check for lingering tests, but if
ran along with our integ tests the test run concluded with...

	Threads lingering after test run:
  	<_MainThread(MainThread, started 140086124480256)>
  	<Thread(Event Notifier, started daemon 140085629839104)>
  	<Thread(Event Notifier, started daemon 140085657794304)>
  	<Thread(Event Notifier, started daemon 140085638756096)>

Tracked this down by dumping the internal attributes of those controllers.
They had mocks from these tests, providing the eureka to get the root cause.