Refactor apply_digest to remove side effects
Per discussion with Damian over IRC, instances of our Cell classes are
intended to be immutable objects. So instead of modifying the RelayCell
and having an additional side effect on the digest, this provides a
modified copy of each, in a tuple.

That makes apply_digest() a pure function, currently with referential
transparency (i.e. also deterministic). In the future, it may utilize
randomness for cell padding, which would make it non-deterministic.

Overall this now more generally follows a functional-programming
paradigm.