Drop unsigned from Size class
Actually, it's already in our pydoc that these are all unsigned. Maybe we'll
add a flag like this in the future if that changes but presently it doesn't
provide any value.

That said, I *do* like the idea of providing a better error message. Keeping
that, and improving our performance a bit by only performing these checks if
packing fails (this class is heavily used in IO so making this a tad extra
performant is useful).