Add check_digest instance method to BaseRelayCell
Similar to how apply_digest() is used by a RELAY cell sender, this
collects the logic needed for a receiver to confirm that a cell is fully
decrypted and that the cell's integrity is intact.

This is new functionality. Prior to this change, stem.client had no
facility to check the digest of a RELAY cell according to the spec.

Not yet used.