Rearrange the digest/key copies to be a bit more spec-compliant
This arrangement is better, as it better keeps track of the digest/key
on a cell-by-cell basis, but it's still not spec-compliant in a number
of ways, such as its potential to drop cells or get corrupted backward
digest/key state in the case of an exception.

That's ok - this is just an interim implementation, and this change is
just to make things a bit better for illustrative purposes.