Add check_recognized_field instance method to BaseRelayCell
This is new functionality. (Nothing in stem.client currently attempts to
check the 'recognized' field.)

Not yet used.

This is arguably a very inefficient method, and it might even obscure
its purpose by existing (it's extremely simple), but it seemed better to
break it out and deal with the foreseeable performance hit than to trend
towards premature optimization.

(It also is a bit counterintuitive for a value of 0 to mean True, so
breaking this out makes that clearer.)

There are many ways this could be optimized, and - especially since so
much of this unit is still in flux - it doesn't make sense yet to do so.