Add interpret_cell instance method to BaseRelayCell
This method facilitates separating decryption / recognition of a cell
from the interpretation of a cell. Many places in the spec indicate that
cells should be dropped (ignored) under certain conditions, but that
such cells should still be accounted for within the running decryption /
digest.

Furthermore, it may be useful in some cases to convert between raw cells
and interpreted cells, without including the other steps of encryption /
decryption.

This additionally allows removing one of the interim TODOs in the RELAY
cell unit tests. Arguably the tests need some rework with all these
changes, but that can come a bit later.

Not yet used, beyond the aforementioned unit tests.

This method will probably evolve a decent amount, since the unit is
still in a state of high flux.