Simplify rollback of key/digest during errors
The reason I asked for our functions to return a new copy of our key/digest is
because it lets us simplify this calling code a bit.

Rather than explicitly reverting the key/digest we can simply wait to record
the new values until cells are successfully sent and received.