Combine relay cell encrypting and packing
Cells are only ever encrypted to be sent. I'm not seeing a purpose in having
an intermediate step here. Our Circuit's send() method is simpler if the
encrypt() method provides us the payload to be sent rather than an intermediate
class.

Are there any use cases where we want an 'encrypted cell' object? We can't do
anything with it besides send it (if we made it) or decrypt it (if sent to us).
Possible though I'm missing something.

Presently I'm only looking at our Circuit's send() method. Our RelayCell's
usage informs the API it should have, so I'm not worrying about the cell module
until that's settled.