Stub minimal cell decryption function
Ok, now into the meat of this sucker. Dave's branch introduces nice additions
such as validating our backward digest and a flag to indicate if we've fully
decrypted the cell or not.

For the moment moving to a minimal decryption function that only supports what
we did previously. Next gonna try to integrate his new features to make a
hybrid approach that hopefully takes the best of both.