Fork new and old RelayCell implementations
Dave's branch makes quite a few RelayCell modifications in support of its new
BaseRelayCell class. That class is no longer used by the new Circuit send
method, but there's quite a few things we want to integrate from it.

Making a total mess, but for the moment placing side-by-side copies of both our
new and old RelayCell class in this module. AlternateRelayCell is Dave's
version, and RelayCell is what we had previously (plus the decrypt stub from
earlier).

This does *not* presently work, as the old RelayCell class lack the encrypt()
and decrypt() methods the Circuit class now excepts. Doing this so we have
a clean starting point to see how we integrate.