New RelayCell encrypt method
"Premature optimization is the root of all evil."

While I love the idea of moving our encrypt() and decrypt() methods into the
RelayCell class I gotta admit I'm not a fan of the verbosity and large number
of helpers. This branche's new RelayCell is far, far more complicated without
adding any new capabilities to the Circuit class.

As such basing this encrypt() method on our previous Circuit send() code.
Hopefully this gets us the best of both worlds: Circuit is simpler because
it no longer does encryption/decryption itself, and our RelayCell remains
an elegantly simple class too.

I suspect Dave might be factoring things this way in anticipation of future
additions but I'd rather not complicate our code until it comes hand-in-hand
with a practical benefit in doing so.