New RelayCell decrypt method
Again, a simple decryption method based on our Circuit's prior send() method.

This branch made a valient attempt to implement our decryption digest. But
honestly the code is twisty enough that I really don't understand how it's
attempting to achieve this (sorry!).

Unless I'm missing something decryption digests only play a role when we're a
middle hop relay, which we're not even close to implementing. This is why
our prior Circuit send() method left it unimplmented.

Maybe I'm mistaken? I'd be delighted to support decryption digests if it serves
a purpose, and is done in a simpler fashion.