Deprecate stem.descriptor.export
This module was provided by a contributor in stem's early days to export
descriptors as csvs. I've never used it, and aside from that contributor
years ago never heard of anyone else using it either.

Lets propose dropping it and see if anyone would mind. Happy to keep it
if it's useful, but if it isn't seems the very definition of bloat. ;)