Integration test race when multiple tor processes are running
Our flakiest integration tests on Jenkins tend to be our system tests...

  ======================================================================
  FAIL: test_pid_by_name_ps_linux
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/require.py", line 58, in wrapped
      return func(self, *args, **kwargs)
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/require.py", line 58, in wrapped
      return func(self, *args, **kwargs)
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/require.py", line 58, in wrapped
      return func(self, *args, **kwargs)
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/integ/util/system.py", line 211, in test_pid_by_name_ps_linux
      self.assertEqual(tor_pid, stem.util.system.pid_by_name(tor_cmd))
  AssertionError: 31963 != None

  ----------------------------------------------------------------------

These tests skip themselves if multiple tor processes are running, *but* they
fail in this fashion if a tor process spawns *during* the test.

To reduce the chances of this the tests now check for additional tor processes
both before *and* after the test. In theory an error could still occure if a
tor process both starts *and* stops in the middle of a test but I'm unsure if
this will really happen in practice. Lets see. If these assertion errors
continue to show up I might simply move them under a special target.