Inconsistent unit test failures in control.controller
On occasion our unit tests that emit then listen for controller events fail
with...

  ======================================================================
  FAIL: test_events_get_received
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/usr/lib/python2.7/dist-packages/mock/mock.py", line 1305, in patched
      return func(*args, **keywargs)
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/unit/control/controller.py", line 684, in test_events_get_received
      self.bw_listener.assert_called_once_with(BW_EVENT)
    File "/usr/lib/python2.7/dist-packages/mock/mock.py", line 947, in assert_called_once_with
      raise AssertionError(msg)
  AssertionError: Expected 'mock' to be called once. Called 0 times.

  ----------------------------------------------------------------------

The specific test varies, but it should always be one of our emit_event()
callers. This has caused nine jenkins test failures in the last three months
(so pretty noisy). Dave ran into this too...

  https://trac.torproject.org/projects/tor/ticket/27173

I can't reliably (or even frequently) reproduce this locally but I think I know
what's going on...

  * emit_event() adds an event to the queue via put(). This is a blocking call
    so we can be assured the event's in there.

  * emit_event() then calls set() on the event_notice condition so the
    Controller's event thread will process the event.

  * The controller's event thread calls get_nowait() on its event queue. This
    *ususally* gets the event we just inserted, but since this is get_nowait()
    rather than get() it's non-blocking, and may raise a queue.Empty instead.

  * When queue.Empty gets thrown we once again wait on the event_notice
    condition.

  * emit_event() then calls close() on its Controller which joins and
    terminates its threads. The close() method calls set() on event_noice to
    give it a *second* opportunity to get the event.

  * Once again, because we use get_nowait() there's a small chance we'll raise
    queue.Empty without processing the event.

  * The event thread then terminates with the enqueued event never being
    processed.

  * Finally, we get the assertion failure above.

This is actually a real bug that can cause Controller users to be highly
delayed or never receive an event. For instance...

  * Tor emits an event.

  * Our reader thread enqueues the event and calls set() on the condition to
    notify the event thread.

  * For the reasons mentioned above it fails to do so.

  * Because we use wait() without a timeout our event thread then resumes
    blocking, causing the event to not get another chance to be delivered until
    a second event comes along.

This sucks, and is pretty easy to remediate. Rather than waiting indefinitely
we should instead have our event thread periodically check re-check its own
queue for undelivered messages (picked 50 ms as a decent sounding balance
between responsiveness and load). Our test's emit_event() in turn now blocks
until the event is delivered.

Needless to say this is a long song and dance so quite possible I'm in the
wrong on something but this seems, both in theory and practice thus far, to be
a good step forward.