New sphinx version shrank download page icons
It wasn't visible on the live site, but when I rebuilt the site on my laptop
(which now runs Sphinx v1.6.5) the download page icons shrank to roughly 80x80.
This is because its haiku css added a new 'max-width: 100%' attribute to
images. Overriding that property to keep things looking how it was.