Digest method for microdescriptors
Revamp digesting of Microdescriptors in the same way as server and extrainfo
descriptors...

  https://trac.torproject.org/projects/tor/ticket/28398

This is definitely better *but* is backward incompatible with the class' old
'digest' attribute. Unfortunately I can't avoid this. The method needs to be
called digest() for consistency, and python cannot have name conflicts between
methods and attributes.

The old digest value was hex rather than base64 encoded which made it
relatively useless (it couldn't be used to fetch or validate microdescriptors,
the sole point of the damn thing) so fingers crossed that no one was using it.

I try very hard to provide backward compatibility in minor version bumps of
Stem but in this case I don't think we should be a slave to that here.