Use newer cached microdescriptors if available
Our data directory has up to two microdescriptor files: cached-microdescs and
cached-microdescs.new.

If the former is unavailable but the later is present we should use it...

  https://trac.torproject.org/projects/tor/ticket/28508

Maybe more important, when looking into this I realized that our attempt to get
tor's data directory stacktraces if not explicitly present in the torrc...

  >>> list(controller.get_microdescriptors())
  Traceback (most recent call last):
    File "<console>", line 1, in <module>
    File "/home/atagar/Desktop/stem/stem/control.py", line 490, in wrapped
      for val in func(self, *args, **kwargs):
    File "/home/atagar/Desktop/stem/stem/control.py", line 1791, in get_microdescriptors
      if not os.path.exists(data_directory):
    File "/usr/lib/python2.7/genericpath.py", line 26, in exists
      os.stat(path)
  TypeError: coercing to Unicode: need string or buffer, NoneType found

Changed it so we'll instead provide a generic exception saying we were unable
to determine the data directory. Fortunatley this whole thing is a fallback, so
eventually we'll be able to remove it.