Separate Query exception attributes
Oops, commit cc43a6c both broke our unit tests and is incorrect in that it
changes the Query's 'error' attribute to a completely different type.

The error attribute is public so we cannot change it. Simply storing exception
metadata as a separate private attribute for the time being. In stem 2.x we'll
be able to drop this whole thing because without requiring python 2.x
compatibility we can re-raise exceptions while retaining their stacktrace.