Reject duplicate bandwidth file entries
Teor adjusted the bandwidth file spec so it is now at the discretion of parsers
for how we behave when bandwidth files list a relay multiple times. My other
descriptor parsers bias toward strictness so doing so here too. Doing so will
hopefully help uncover if we ever get sbws regressions, and we can always
adjust this back to something looser later.