Loosen 'tor --version' parsing
Nick formally documented the 'tor --version' output we rely on (thanks!)...

  https://gitweb.torproject.org/tor.git/commit/?id=0e6e902

Stem asserted the version is the last line of the output but that is not
specified, and I can see scenarios where we might add more text. Loosening our
parser to look for the version in any line of the output and expanding our
tests.