Skip autocompletion for non-interactive interpreter
Sadly I forget where it was pointed out, but invoking the control port via
shell is a *lot* faster than stem...

  #!/bin/bash -e

  cmd="$@"
  pass="ControlPortPassword"

  function test_tor() {
      echo "$1" >&3
      sed "/^250 OK\r$/q" <&3
      echo QUIT >&3
      exec 3<&-
  }

  exec 3<>/dev/tcp/127.0.0.1/9051
  echo AUTHENTICATE \"$pass\" >&3
  read -u 3
  test_tor "$cmd"

  ====================

  atagar@morrigan:~$ time ./bench.sh 'GETINFO version' 1>/dev/null

  real	0m0.007s
  user	0m0.004s
  sys	0m0.003s

  ====================

  atagar@morrigan:~$ time tor-prompt --run 'GETINFO version' 1>/dev/null

  real	0m0.186s
  user	0m0.072s
  sys	0m0.030s

Generally speaking this is expected. Spinning up an interpreter takes time. But
in doing a quick investigation realized this is quite a bit slower than it
needs to be...

  total tor-prompt runtime          0.186 seconds
  --------------------------------------------------------
  python interpreter startup        0.016 seconds (9%)
  import statements                 0.079 seconds (42%)
  check if tor is running           0.014 seconds (8%)
  connect to tor                    0.009 seconds (5%)
  autocompete setup                 0.065 seconds (34%)
  invoke tor controller command     0.003 seconds (2%)

Autocompletion is only relevant when the user is presented with an interactive
interpreter. If we're merely invoking a command it's pointless.

So TL;DR: tor-prompt is now ~34% faster when used to invoke controller commands.