Bandwidth file version must be second header
Good point from juga that the bandwidth file spec requires the 'version' header
to be in the second position (unless the document is version 1.0.0, in which
case the version attribute is not present at all).

  https://trac.torproject.org/projects/tor/ticket/29539

Ensuring the descriptors we create conform with this, and if the user parses a
bandwidth file with 'validate = True' then check this.