Note thread names rather than idents
Thread ident numbers are unhelpful for troubleshooting hangs. If a thread hangs
I need to know its name, not its internal numeric identifier.

Also cleaning up the pid bits a tad. The pgid variable was unused, and there's
no need to pass the pid of ourselves as an argument (it's static for the
interpreter, so we can fetch it as needed).