Fix ExitPolicy thread safety bug
Interesting catch from juga! I don't have a repro for the stacktrace they cite,
but the only way our input_rules could potentially be None is if two threads
attempt to parse the input_rules at the same time.

Creating a local reference for the input_rules to avoid this while remaining
lock-free...

  https://trac.torproject.org/projects/tor/ticket/29899