Support multiple '--test' arguments
When supplied our '--test' argument limits us to only running that specific
test. I added this to help with single-test troubleshooting, but on reflection
it could be useful to the network team as well to limit their CI to only run
the tests they're concerned with.

Only gotcha was that we didn't support multiple '--test' arguments. Resolving
that.