Move compression argument into index method
Originally I thought that CollecTor offered each file with all compression
types (like DirPorts) but on reflection that's wrong. The index is the only
file that's compressed in multiple ways so moving the 'compression' argument
from CollecTor's constructor to the index() method.

Overall this makes things quite a bit cleaner.