Support a DocumentHandler when reading consensus documents
When reading a consensus our callers need the ability to specify what form they
wish to receive the document (just the document, a document with the relays
populated, or just the relays).

Unfortunately CollecTor has separate annotations for v3 consensus and vote
documents, whereas all other classes have at most one. This requires rethinking
the annotation constants but for the moment moving on to bigger fish.