Adjust certificate validate() TODO comment
George is absolutely right that this method is too specific to server
descriptors, but I cannot find the generic "validate()" method he's
referring to. Just reached out to him on irc to discuss this.

If we have a generic method that applies to both server and hidden
service descriptors I'd be delighted to deprecate this, but for the
moment at least I'm unsure how to proceed with this.