Parse signing_cert as an ed25519 certificate
I was mistaken to parse signing_cert as a key blob. Server descriptors are
deprecating their key blob copy (the 'ed25519_certificate' attribute) in favor
of a Ed25519Certificate instance, and hidden service descriptors should do the
same.

Hidden service v3 parsing isn't yet in a release so we can simply fix this in
place without concern for backward compatibility.