Add reference implementation for ed25519 key blinding.
This is straight from little-t-tor (see #31912).

We need this specific implementation because we can't do v3 key blinding with
hazmat or pynacl, since they don't expose the primitives we need. So we are
gonna use this simple ed25519 implementation to do key blinding when
needed (it's very rarely needed so it's not affecting performance).