Expose classes for HSv3 blinded keys.
HSv3 blinded keys are not compatible with hazmat keys because their private key
format is different. This means that we cannot store them in a hazmat or pynacl
data structure.

This commit adds an abstraction layer on top of the ed25519 blinded keys that
implement the fundamental methods of the hazmat ed25519 module (duck typing),
so that they can be used interchangeably to sign and verify documents.