Remove _is_crypto_ed25519_supported() helper
We have a couple checks that are unpleasantly similar...

  * "is_crypto_available(ed25519 = True)" checks if the cryptography module
    has ed25519 support (ie. if cryptograhy itself is modern enough)

  * "_is_crypto_ed25519_supported()" checks if OpenSSL has ed25519 support

This is confusing, so combining these into a single "is ed25519 fully
supported?" check. This potentially makes our checks stricter (ie. we do
less when one check is true, and the other is false) but this isn't worth
the additional confusion.