Reduce Ed25519CertificateV1 constructor arguments
Why did I have a 'version' argument? The version is part of the class name - if
it ever *isn't* one that's a problem. ;P

Also, having separate encoded/decoded constructor arguments for the exact same
thing was silly. Decoding base64 twice is cheap, and having two arguments risks
having them mismatch (which would be a frustrating headache to troubleshoot).