HSv3 descriptor creation test
Now that we've revised HiddenServiceDescriptorV3.content() we can test it
in a similar fashion to inner/outer layers. This provides the same coverage
as test_encode_decode_descriptor() so that test is now redundant.

Fixing a few bugs this spotted along the way and renaming the
'public_key_from_address' functions to specify that they take identity keys
instead (we have so many key types I lost track of what these truly did when I
originally tacked 'em on).