Disable key blinding by default
On reflection, why not simply disable blinding by default? Blinding parameters
are unnecessary for the vast majority of use cases, and when needed the caller
can simply provide a nonce (and by extension accept the lengthy runtime).

We can always re-enable blinding as a default if/when we have a performant
implementation.