Require Python 3.6 to run tests
Now that the PSF has removed Python 2.x support we are as well...

  https://www.python.org/doc/sunset-python-2/

Stem 2.x will require Python 3.x, but what version? Python 3.5 is EOL in a few
months (prior to our next release)...

  https://devguide.python.org/#status-of-python-branches

So the question is 3.6 or 3.7? I'm not yet aware of a reason to pick 3.7 so for
the moment picking 3.6. This might change if we strongly desire a 3.7 feature
as we work on this branch.

  https://docs.python.org/3/whatsnew/3.7.html