Don't suppress socket closing exceptions
We'll revert this if it becomes an issue in practice, but the comment says that
we suppress exceptions due to python 2.x. I tried closing a socket multiple
times with python 3.x and it did not raise an exception, so for the moment we
don't have a known reason to expect these calls to error.