Drop BytesIO TODO comment
BytesIO is great to concatenate large data or turn strings into file objects,
but I'm kinda unsure why I wanted one here. This buffer both adds *and* pop off
data, and BytesIO are kinda clunky at the later. Perhaps there's a good reason
I'm forgetting and we'll do so in the end, but for now just dropping the TODO.