Differentiate bytes/unicode in hashes
Huh. To avoid bulk refactoring our tests we included a small hack that hashed
byte and unicode types to the same value. Said another way, a unicde and byte
string for "hello" would be equal.

I expected that dropping this hack would be invasive but our tests pass. Kinda
confused, but... neat?