Drop server descriptor annotations
Server descriptor files within tor's data directory begin with a couple
annotations. This is the only place where these annotations exist, and
is not part of tor's dir-spec.

I parsed these annotations for completeness when I first authored Stem.
However, we effectively never read these files nowadays for a few reasons...

  1. Stem's remote and collector modules provide better methods for getting
     descriptors.

  2. When tor is running locally we can get this information via GETINFO rather
     than reading directly from disk.

  3. Tor is trying to discourage us from reading the data directory from disk
     by restricting its permissions.

This isn't to say there aren't merits to reading on-disk descriptors. Nyx does
so, for instance, because it's a passive listener (no downloading allowed)
and massive GETINFO responses noticably lag the control port.

However, reading on-disk descriptors is rare and even in this edge case I'm
unaware of anyone using the annotations.