Drop undocumented 'encoded' attribute from certificates
When an Ed25519Certificate was created through from_base64() we appended an
undocumented 'encoded' attribute with the base64 content we were created from.
Our to_base64() method computes the exact same thing.

This attribute was only used in a single place for server descriptor
validation.

If 'encoded' and to_base64() mismatches that is a bug. Fiddling with these I
cound't come up with a scenario where that is a case, so dropping the redundant
attribute.

Also merging _validate_server_desc_signing_key() into its sole caller.