Prototype Controller asyncio usage
Still very much an experiment. Several things are broken but this adjusts our
Controller successfully wrap an asyncio socket.

  from stem.control import Controller

  controller = Controller.from_port()
  controller.connect()
  controller.authenticate()
  print("Tor is running version %s" % controller.get_version())
  controller.close()

  ----------------------------------------

  % python3.7 demo.py
  Tor is running version 0.4.3.0-alpha-dev (git-5e70c27e8560ac18)

Essentially this replaces our reader daemon with an asyncio loop. Synchronous
code can invoke this loop in the following ways...

  * Non-blocking call of a coroutine

    self._asyncio_loop.create_task(my_coroutine())

  * Blocking call of a coroutine

    result = asyncio.run_coroutine_threadsafe(my_coroutine(), self._asyncio_loop).result(timeout)

  * Non-blocking call of a function

    self._asyncio_loop.call_soon_threadsafe(my_function)

I'm not quite sold yet on if we'll go this direction. An asynchronous socket
has two potential advantages...

  * Simplify multi-threading
  * Offer asynchronous variants of our methods

Either or both of these might be dead ends. I'll need to experiment more, but
now that we've seen this work in this limited way we have something to build
off of.