Match loop scope to thread
Asyncio threads can be restarted, but doing so lacks a significant benefit and
can get complicated. For instance, when we're stopped from an async method our
loop is closed asynchronously (because we cannot join our own thread). This is
fine, except that start() can subsiquently fail because we cannot resume a
running loop...

  Traceback (most recent call last):
    File "/home/atagar/Python-3.7.0/Lib/threading.py", line 917, in _bootstrap_inner
      self.run()
    File "/home/atagar/Python-3.7.0/Lib/threading.py", line 865, in run
      self._target(*self._args, **self._kwargs)
    File "/home/atagar/Python-3.7.0/Lib/asyncio/base_events.py", line 510, in run_forever
      raise RuntimeError('This event loop is already running')
  RuntimeError: This event loop is already running

By creating a new loop for each thread we not only sidestep this but simplify
asynchronicity beacause each run of our class will have its own event queue.