Resume Synchronous when async methods are called
When our Synchronous class was stopped all further invocations of an async
method raised a RuntimeError. For most classes (sockets, threads, etc) this is
proper, but it made working with these objects within synchronous contexts
error prone.

For example, our Controller's async connect() method resumes our instance, but
was uncallable due to this behavior. Stopping should be the last action callers
take, and failing to so so is inconsequential (it simply orphans a daemon
thread) so erring toward our object always being callable.