Drop usage of get_running_loop()
Stem supports Python 3.6+ but asyncio.get_running_loop() was added in 3.7.

The only difference between get_running_loop() and get_event_loop() is that the
former raises a RuntimeError when outside an asyncio context. These calls are
assured to be asyncronous so it really doesn't matter - I just picked
get_running_loop() because it reads a tad better.

The one exception is our is_asyncio_context() function. For that we need a
fallback, which fortunately a private method provides...

  https://github.com/Azure/msrest-for-python/issues/136