Move controller connecting to aenter
Static methods such as from_port() and from_socket_file() cannot invoke
asynchronous methods. Fundimentally this is the same problem as our ainit -
when a loop is transtively running us we cannot join any futures we create.

Luckily in this case we can simply sidestep the headache. from_port() and
from_socket_file() are designed for 'with' statements so we can simply move the
act of connecting into our context management (which is already asynchronous).

I encountered this problem when I ran the following...

  import asyncio

  from stem.control import Controller

  async def print_version_async():
    async with Controller.from_port() as controller:
      await controller.authenticate()
      print('[with asyncio] tor is version %s' % await controller.get_version())

  def print_version_sync():
    with Controller.from_port() as controller:
      controller.authenticate()
      print('[without asyncio] tor is version %s' % controller.get_version())

  print_version_sync()
  asyncio.run(print_version_async())

Before:

  % python3.7 demo.py
  [without asyncio] tor is version 0.4.5.0-alpha-dev (git-9d922b8eaae54242)
  /home/atagar/Desktop/stem/stem/control.py:1059: RuntimeWarning: coroutine 'BaseController.connect' was never awaited
    controller.connect()
  [with asyncio] tor is version 0.4.5.0-alpha-dev (git-9d922b8eaae54242)

After:

  % python3.7 demo.py
  [without asyncio] tor is version 0.4.5.0-alpha-dev (git-9d922b8eaae54242)
  [with asyncio] tor is version 0.4.5.0-alpha-dev (git-9d922b8eaae54242)